+++ checkpoint/CHANGELOG.md
@@ -13,6 +13,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Changed
 - Explicit Python version support for 3.9, 3.10, 3.11
+- Raise ValueError when trying to save jax.Array to the aggregate file if it is
+not fully replicated.
+
 
 ## [0.2.3] - 2022-05-12